examples: Add udp-pingpong application
This patch introduces a trivial udp-based application which tests several
platform-specfic systems such as console output, leds, button-sensor and
networking.

The application works as follows. When the user press a button, the node
'A' sends a udp packet which represents a "token". Node 'B' receives the
token, prints a message in console and turns on the LED. After 3 seconds
node 'B' sends the token and turns off the LED. Once node 'A' receives
the token it does the same: print a message, turn on the LED, wait 3
seconds and sends the token back to node 'B'. The token goes back and
forth until one of the nodes is turned off.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>