galileo: Add GPIO interrupt example
This patch introduces an example application to demonstrate how to use
GPIO driver APIs to manipulate interrupt pins.

The application uses default galileo pinmux initialization and sets
the GPIO 5 (IO2) as output pin and GPIO 6 (IO3) as interrupt. It toggles
the output pin stat at every half second in order to emulate an interrupt.
This triggers an interrupt and the application callback is called.