quark-d2000-devboard: Add support for Clock module
This patch adds support for Contiki's clock module. All functions from
core/sys/clock.h are implemented, except clock_set_seconds() and clock_
delay_usec().

The CLOCK_CONF_SECOND macro is set to 128. This value seems to be good
enough since several platforms use it.

We use the PIC timer in periodic mode to implement the Clock module so
this patch also fixes the README.md to list the PIC timer as one of the
devices already used the OS.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>
Signed-off-by: Jesus Sanchez-Palencia <jesus.sanchez-palencia@intel.com>