quark-se-devboard: Add support for LED APIs
This patch implements LED architecture-specific functions in order
to enable LED APIs for Quark SE Devboard platform.

This board has two "user" LEDs. LED 0 is connected to GPIO PIN 25
and LED 1 is connected to GPIO PIN 26. The LED is turned on when
GPIO pin is set to '0' and turned off when set to '1'.

Finally, Contiki API identifies LEDs by color. In this implementation,
LED 0 is LEDS_GREEN and LED 1 is LEDS_YELLOW.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>