quark-se-devboard: Add cc2520 driver
Add a cc2520 driver for the AtlasHills platform.
This implementation is based on the original dev/cc2520/ module,
but with some modifications to the cc2520 implementation and overall
structure.

The idea here is basically to keep the original cc2520 API from
dev/cc2520/cc2520.h, but making the macros call platform hooks
defined by cc2520-arch.h instead of using core/dev/spi.h macros.
These platform hooks are implemented at cc2520-arch.c .

In fact, we shouldn't need the CC2520_* macros at all as the
cc2520 functions implementation could be calling the platform
hooks directly. However, for now we are bound to whichever APIs are
exposed by the original dev/cc2520/cc2520.h interface and those
macros are part of it.

The approach followed here can be merged into the original module,
making it easier for a platform to implement the platform hooks needed.

Signed-off-by: Jesus Sanchez-Palencia <jesus.sanchez-palencia@intel.com>