quark-se-devboard: Use whole flash memory
This patch adds a patch to QMSI which is applied before we build the
QMSI BSP. The patch changes the default ROM firmware so it jumps to
the beginning of flash0 bank instead of flash1. It also modifies the
application linker script so the 'flash' region begins at flash0.

The quark-se-devboard doesn't support the ARC sensor subsystem (use
is the primary user of flash0 bank). So this change enable us to use
the whole flash memory for Contiki applications.

To this patch takes effect, you should first flash the Contiki
application and then the ROM firmware.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>