qmsi-common: Add support for _exit syscall
This patch adds the newlib-syscalls-extra.c which implements the system
calls that are not already provided by QMSI BSP. Since these system calls
can be shared by any QMSI-based platform, we land this .c in qmsi-common.

For now, we are aware of only one system call (_exit) which is required
by soletta framework [1]. We might add support for more system calls in
future according to application requirements.

[1] https://solettaproject.org/

Signed-off-by: Andre Guedes <andre.guedes@intel.com>