x86: Define AROPTS in Makefile.x86_common
In Makefile.include, if variable 'AROPTS' is not defined, the options
"rcf" are used by default. The option "f" truncates names in the archive
which is a bit annoying when we are inspecting the contiki-$TARGET.a
artifact. See an example below:

$ ar -t examples/hello-world/contiki-quark-se-devboard.a
...
uip-debug.o
uip-packetqueue
uip-udp-packet.
resolv.o
uip-nameserver.
frame802154e-ie
...

According to 'man ar', the "f" option truncates names in order to keep
compatibility with 'ar' from some systems. This doesn't look like a
concern to x86-based ports since 'ar' from x86 toolchains deals with
long names well.

That said, this patch changes Makefile.x86_common so we define our own
'AROPTS' without "f" option. It also adds the "s" option so the
object-file index is build. Just for reference, our AROPTS are the same
used to build libqmsi.a.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>