qmsi-common: Add .PRECIOUS to Makefile.customrules-qmsi-common
The files %.co and $(PROJECT_OBJECTFILES) are treated as intermediate
files and are automatically deleted by 'make' after the final image is
created. See the output of hello-world examples:

$ make TARGET=quark-se-devboard -C examples/hello-world
...
  AR        contiki-quark-se-devboard.a
  CC        hello-world.c
  CC        ../../platform/qmsi-common/newlib-syscalls-extra.c
  LD        hello-world.quark-se-devboard
   text	   data	    bss	    dec	    hex	filename
  83976	   1156	  11496	  96628	  17974	hello-world.quark-se-devboard.elf
rm obj_quark-se-devboard/newlib-syscalls-extra.o hello-world.co
make: Leaving directory '/mnt/extra/src/contiki-qmi/examples/hello-world'

For this reason, two consecutive builds of a given application always
build those files again. This is particular annoying when we flash
an application built with BUILD_RELEASE=1. In this case, we must set
BUILD_RELEASE=1 in the flashing command line ('make flash'), even if
it doesn't make much sense, otherwise %.co and $(PROJECT_OBJECTFILES)
are not build in "release mode".

To fix this, we add those files to .PRECIOUS dependency list. This
way they are not treat as intermediate files and are not deleted
automatically. This is the same approach used in Makefile.include
to avoid %.$(TARGET) being deleted, by the way.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>