qmsi-common: Fix custom linking rule
Two consecutive builds from the same application always execute the
linking rules, even if no files are changed. This happens because
our custom linking rule doesn't generate the target, in this case
%.$(TARGET).

So this patch changes the custom linking rule so it generates the
target. Besides fixing the issue with two consecutive build reported
above, we keep the name convention from the final image consistent
with other ports (the convention is $(CONTIKI_PROJECT).$(TARGET).

Signed-off-by: Andre Guedes <andre.guedes@intel.com>