quark-se-devboard: Add support for button-sensor API
This patch implements the basic support for button-sensor API for Quark
SE Devboard.

The button-sensor API lacks documentation so the returning code
convention from configure() callback is based on the convention used
by other platforms such as sky, z1 and cc2530dk: '1' for success and
'0' for error.

In this intial version the callbacks value() and status() are not
implemented since, due to the lack of documentation, it is not clear
what we are supposed to do, and it seems they are not required to
have the basic button-sensor systems working.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>