quark-se-devboard: Remove all #ifndef occurrences in contiki-conf.h
The contiki-conf.h has several occurrences following the format:

 #ifndef SOME_MACRO
 #define SOME_MACRO
 #endif

This patch removes all of them since they are not that useful and
"pollutes" the contiki-conf.h file, making it a bit painful to
follow.

Moreover, the macro UIP_CONF_BUFFER_SIZE is defined twice so we remove
the second occurrence.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>